requirements_new.txt
absl-py==1.2.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
distlib==0.3.6
filelock==3.8.0
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
oauthlib==3.2.1
platformdirs==2.5.2
protobuf==3.19.4
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pywebio==1.6.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scienceworld==1.0.2
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.91
six==1.16.0
sklearn==0.0.post5
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.8
threadpoolctl==3.1.0
torch==1.8.1
tornado==6.2
typing_extensions==4.4.0
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
virtualenv==20.16.5
Werkzeug==2.2.2
zipp==3.9.0
